The library can now be shaded instead of be accessed as a plugin
